feat(backend): estabilização de arquitetura, expansão de testes e auditoria documental

- [KAN-30] Fix: Correção de 9 erros de compilação pós-refatoração, substituindo as chamadas obsoletas a `Alert.timestamp` por `Alert.createdAt` no Mapper e nos Serviços de Relatório.
- [KAN-30] Fix: Correção de queries nativas em `RuleService` (`findAllByActiveTrue` removido) e adição do campo `conditions` ao DTO `RuleCondition`.
- [KAN-31] Test: Reparação de `NullPointerExceptions` em `SimulationServiceTest` injetando `@Mock SimulationEngineService`.
- [KAN-31] Test: Correção do porto de verificação do `EmailServiceTest` (5174 -> 5173).
- [KAN-31] Test: Criação de testes unitários robustos para `BioGearsParserService` simulando payloads com falhas.
- [KAN-31] Test: Adição de cobertura explícita no `AuthServiceTest` para documentar a blindagem contra Soft Deletes nativos (`@SQLRestriction`).
- [KAN-32] Docs: Auditoria global de código, tradução de comentários (PT -> EN) e padronização rigorosa com anotações Javadoc nos recém-criados Controladores e Serviços.

+++ src/main/java/com/grupo3aor/innovationlab/dto/AlertDTO.java
@@ -3,7 +3,7 @@ package com.grupo3aor.innovationlab.dto;
 import com.grupo3aor.innovationlab.domain.enums.AlertStatus;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-import java.math.BigDecimal;
+
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -27,5 +27,5 @@ public class AlertDTO {
     private AlertStatus status;
 
     @NotNull(message = "The reading value recorded at trigger time is required")
-    private BigDecimal valueAtTrigger;
+    private Double valueAtTrigger;
 }
